requirements.txt
absl-py==2.0.0
annotated-types==0.6.0
anyio==3.7.1
astunparse==1.6.3
beautifulsoup4==4.12.2
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
deepface==0.0.79
distro==1.8.0
dnspython==2.4.2
exceptiongroup==1.1.3
filelock==3.13.1
fire==0.5.0
Flask==1.1.2
flatbuffers==23.5.26
gast==0.5.4
gdown==4.7.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
grpcio==1.59.2
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.2
httpx==0.25.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
keras==2.15.0
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==1.1.1
ml-dtypes==0.2.0
mtcnn==0.1.1
numpy==1.26.2
oauthlib==3.1.0
openai==1.3.0
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
protobuf==4.23.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==2.5.1
pydantic_core==2.14.3
pymongo==4.6.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.25.1
requests-oauthlib==1.3.0
retina-face==0.0.13
rsa==4.7.1
six==1.15.0
sniffio==1.3.0
soupsieve==2.5
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.3
Werkzeug==1.0.1
wrapt==1.14.1
